pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cafa5-protein-function-prediction"
version = "0.1.0"
description = "CAFA-5 Protein Function Prediction using protein language model embeddings"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "tqdm>=4.65",
    "torchmetrics>=1.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"
